switch simon says and ajaxifying hn on Tuesday
cc: @sgharms @mikelikesbikes @mattbaker @caseee 

Students have responded negatively to the current challenge order in the past, reporting that by the time they got started on AJAXifying Hacker News, it was well after core hours, and they had no staff or mentors around to help them get unstuck - particularly on Release 0, which establishes the first major milestone that can hold the rest of the functionality together in future releases. 

This week, I messaged to students that the challenge order was not set in stone, and they might choose to pull AJAX-HN forward in the day's workflow if they had found themselves exhausted and defeated by final cores on previous days. One student left very positive feedback in response to this messaging.

I predict switching the last two W5 Tuesday challenges would give the day more of a "warm up, get sweaty, cool down" feel, and let instructors provide more direct feedback and guidance on the day's meatiest challenge before we've left for the day.