Add Critical Convos / Team Dynamics EE to 1.1.5
Phase 2 is the first that heavily leans on team
work.

Let's introduce the ability to bring up conflict.
I hope we're going to get the difficult
conversations idea, the team agreements idea
(let's not work unto death), etc. If we can get
this on a good keel, we're mitigating possibility
of explosion in the microagg EE in the subsequent
week.