Add Microgressions EE to 2.2.3
Adding on Wednesday to create a solid distance
from the Assessment. This EE, when facilitated
poorly or with the wrong mindset dominant in the
room can go horribly wrong and, I think, affect
students' performance in the assessment. I want to
create a big gap here.

Also, this requires a 90 minutes video watch. I
want to create several days of watching and
chewing of the prep material in small bites.